hw/hw6/backend/requirements.txt
Flask==2.2.2
google-cloud-secret-manager==2.10.0
functions-framework==3.*
